refactor: organize project structure
