+++ client/src/components/ChatContent.vue
