+++ client/src/components/ChatLayout.vue
