+++ client/src/components/ServerSidebar.vue
