+++ client/vue.config.js
