+++ server/server.js
@@ -64,11 +64,11 @@ console.log(`Current NODE_ENV: "${process.env.NODE_ENV}"`);
 if (process.env.NODE_ENV !== 'production') {
   console.log('Running in development mode - serving frontend files');
   // Serve static files from the 'dist' folder (production build)
-  app.use(express.static(path.join(__dirname, 'dist')));
+  app.use(express.static(path.join(__dirname, '..', 'dist')));
 
   // Return the main index.html for all routes (SPA)
   app.get('*', (req, res) => {
-    res.sendFile(path.join(__dirname, 'dist', 'index.html'));
+    res.sendFile(path.join(__dirname, '..', 'dist', 'index.html'));
   });
 } else {
   console.log('Running in production mode - not serving frontend files');
